Moved FrontPage component to subfolder with Login and SignUp components
