+++ client/src/components/FrontPage/Login.jsx
@@ -1,0 +1,7 @@
+
+function Login() {
+
+
+}
+
+export default Login
