+++ client/src/components/FrontPage/SignUp.jsx
@@ -1,0 +1,6 @@
+
+function SignUp() {
+
+}
+
+export default SignUp
